Refactored Tk infixed UI classes to not inherit from a more general UI class
The behavior of the base class was too minimal to justify the extra complexity added by the level of inheritance. I also realized that it is unlikely that two kinds of user interfaces will be active developed at the same time. The browser-based app will probably be a separate program, while the possible switch to PyQT for the desktop will substitute, and not be concurrent, to the tkinter interface.