Tkinter's main loop now runs on main thread. Other threads are now daemonic.
This can always be reversed if Tkinter needs to be called from a separate thread for testings purposes. However, I don't see why we could not call the main loop from a different thread inside the tests, if that was needed. I struggled with this a lot in a past version of the program and I don't quite remember what alternatives I've tried, so maybe reverting will turn out to be the only way out, after all. The remaining threads should be left daemonic. Maybe some protection against corrupting autosaves should be added as daemon threads exit abruptly.