[MAJOR DESIGN OVERHAUL] For more information, see commit message.
Major design overhaul. This significantly reduces coupling in various parts of the code, most notably between the logic and UI. Changing the UI engine should be more feasible now. Readability should by greatly improved in the higher-level classes, as well. The main changes are:
- `TimelineCollection` class renamed to `Timelines`
- `TimelineUICollection` class renamed to `TimelineUI`
- `TiLiA` class renamed to `App`
- Decoupled backend from frontend:
	- `Timelines` from `TimelineUIs`
	- `Components` from `Elements`
	- `App` from `Timelines`
	- Backend and frontend now communicate through the request system
- `App` dependencies are now passed as arguments to its constructor, instead of being created inside it
- Increased cohesion in `App` class:
	- Moved all purely file related functionality out of `App` and into `FileManager`
	- Moved Autosaving functionality into `AutosaveManager` class
	- Moved app logic for loading media out of `App` into a `MediaLoader` class, which is instantiated only when loading media
	- `TimelineUICollection` was moved to the `UI` controller
- Overhaul in the event system, now called request system:
	- Events that don't require a response are now in the `Post` class. Those may be *listened* for by listeners. This corresponds to the previous publish/subscribe logic.
	- Events that require a response are in the `Get` class. Users may request data using the `get` functions. For responding, callbacks are registered using the `serve` function.
- `Timelines` are decoupled from `TimelineUIs`, they communicate through
- Moved initializing logic out of `main.py` into dedicated `boot.py` module
- `Flake8` used for linting the code

Along with those, numerous smaller refactors were performed. Some new tests were also added, Ideally, the changes is this commit would be split into dozens of smaller commits, but I believe the time investment would not be justified, asthe chances are pretty low that someone will need to look into the code before this refactor. Let's hope that is the case. If you, in the future, is trying to understand the changes in this commit, my sincerest apologies.

Instability is to be expected until more thorough use-testing is performed. The plan is to supplement the existing tests with ones that prevent the regression of bugs that show up in the use tests. THIS SHOULD BE DONE BEFORE THE NEXT RELEASE.