Fixed YT music videos not loading
Turns out, the problem was a missing HTTP header. I would never have figured this out, so we're eternally indebted to this SO user: https://stackoverflow.com/a/76507387.

I have tried out 10 videos that we're not loading before this commit, and 9 of them are now currently working. The one that does not displays a different, and helpful, message:

"Playback on other sites has been disabled by the video owner"

No way of getting around that, so let's hope video owners don't get into that habit.