Fix accidental display when using MusAnalysis font
The flat char is mostly fixed, but I learnt that the font does not support double sharps or flats, so I had to hack something to cover those cases. I couldn't make the double sharpt symbol readable, so I resorted to to using to consecutive sharps. I also wasn't able to get the flat symbol before a roman numeral not not be superscript. Is that by design? I feel like a lot of analysts are going to be bothered by it.

Closes #121