Update applied_to harmony attributes when creating or deleting modes
This is necessary since what is stored by the harmonies is the step that they apply to, and not the interval between the harmony step and the "target" step. That was bad design on my part, as it is the latter that is kept constant through mode changes. We should change that in the future, and maybe follow what [ChordHand](https://github.com/FelipeDefensor/ChordHand) does.