Refactor FileManager.open and improve error handling
I feel that the App class is better suited to deal with the task, as
it has direct access to the app state and to the function to actually
load the file.

The initial motivator for this change was the need to pass a path to
post(Post.FILE_OPEN), as the CLI will need that for a non-interactive
open command. Only after some time I realized there was already
another Post value that triggered a FileManager method which could take
the path as an argument and did not request it from the user. Anyway, the
refactor results in more concise code as both functionalities now use the
same Post value.

The refactor also made it easy to register recent files when opening, and
not when saving, which I feel is what users expect. I found myself
several times saving unmodified files only so they would appear in the
recent files menu later.