Refactor timeline creation
Timeline subclasses now create their own component managers, which simplifies the creation process. I was going with dependency inversion and passing the component managers as arguments. There is no reason to avoid coupling the component manager to the timeline, however, as I can't imagine a scenario where we would want to use a different component manager than the usual one.