Let TiliaActions specify arguments
Arguments supplied to ActionParams will now finally be passed along when posting the action's request. This allows us to use the same action for multiple related purposes (see next commit for an example). A lot of code can now be deduplicated, such as that related to window opening/closing and timeline creation.