Abstract Post.TIMELINE_ADD_X
As we can now pass arguments when using `TiliaAction`, we can use a single Post.TIMELINE_ADD and pass the timeline kind as an argument. That enables us to remove a lot of code duplication and also make creating new TimelineUI subclasses easier, as there are now fewer places that need to be changed.