Move SvgViewer to frontend
SvgViewer displays data to the user, and does not store, so it shouldn't be in the backend. In addition, it depends on the QtUI frontend specificlly, so requiring it in the backend would probably cause crashes in the CLI.

We can move SvgViewer to the frontend and use requests to communicate with it when needed.