Remove post_process.py
Removing the post-processing logic simplifies the on_*_request methods considerably.

I think that post-processing requests was a premature abstraction. If we ever need further computation that depends on the results of function calls to multiple timelines we can just do that in the function that handles the request in TimelineUIs.

The only request that was actually relying on post-processing was Post.TIMELINE_ELEMENT_COPY, so that had to be refactored. I took the chance to (not prematurely) abstract it, removing a lot of duplicated code.