Make RequestHandler methods return type consistent
Require a boolean to be returned from `RequestHandler` handlers. The old approach with the `fallible` decorator and the `RequestFailure` was being used inconsistently and seemed to me less clear than simply returning booleans.

`TimelineUIs.validate_success_return_type` will print a warning (but not display it to the user) if some method does not return a boolean. We don't want to alarm the user here since a wrong return type will only interfere in whether the action gets recorded by `UndoManager`.

This process also made a lot of undo/redo bugs apparent. For instance, marker, beat and harmony creation were still being recorded even when the operation failed (e.g. if there was already a component at the same time). Those were fixed.