feat: make export more JSON-like
We were using lists instead of objects to store components to reduce duplication in the JSON file. However, if file size is important, as in a web context, compression takes care of the object key duplication. Using object has the benefit of making the output format, less complex, less surprising and easier to document.