fix: returns before event loops complete
A tiny patch on one of the bigger architectural issues - many of Qt's modules run on event loops and we don't tend to run our end as an event driven system. This fixes some of it by waiting for Qt to respond with the corresponding signal before returning control to the user.